Fixed BirdInfor and docker file
- Seperated bird info into its own class and endpoint
- Changed the docker compose file to have a healthcheck to check that
the server is running
