+++ docker-compose.yml
@@ -1,20 +1,6 @@
 version: '3.8'
 
 services:
-  webapi:
-    container_name: webapi
-    build:
-      context: .
-      dockerfile: Dockerfile
-    ports:
-      - "5000:8080"
-    depends_on:
-      - db
-    environment:
-      - ASPNETCORE_ENVIRONMENT=Development
-      - ConnectionStrings__DefaultConnection=Server=db;Database=BeakPeek;User=sa;Password=Your_password123;TrustServerCertificate=true;
-    volumes:
-      - ./res/testData:/data
   db:
     image: "mcr.microsoft.com/mssql/server:2022-latest"
     container_name: sql_server2022
@@ -24,4 +10,26 @@ services:
       SA_PASSWORD: "Your_password123"
       ACCEPT_EULA: "Y"
     restart: always
+    healthcheck:
+      test: /opt/mssql-tools/bin/sqlcmd -S localhost -U sa -P "Your_password123" -Q "SELECT 1" -b -o /dev/null
+      interval: 10s
+      timeout: 3s
+      retries: 10
+      start_period: 10s
 
+  webapi:
+    container_name: webapi
+    build:
+      context: .
+      dockerfile: Dockerfile
+    depends_on:
+      db:
+        condition: service_healthy
+        restart: true
+    ports:
+      - "5000:8080"
+    environment:
+      - ASPNETCORE_ENVIRONMENT=Development
+      - ConnectionStrings__DefaultConnection=Server=db;Database=BeakPeek;User=sa;Password=Your_password123;TrustServerCertificate=true;
+    volumes:
+      - ./res/testData:/data
